@@ -27,13 +27,11 @@ namespace WalletWasabi.Backend
 
 		public P2pNode P2pNode { get; private set; }
 
-		public BlockNotifier BlockNotifier { get; private set; }
+		public HostedServices HostedServices { get; set; }
 
 		public IndexBuilderService IndexBuilderService { get; private set; }
 
 		public Coordinator Coordinator { get; private set; }
-
-		public ConfigWatcher RoundConfigWatcher { get; private set; }
 
 		public Config Config { get; private set; }
 
@@ -42,6 +40,7 @@ namespace WalletWasabi.Backend
 		public Global(string dataDir)
 		{
 			DataDir = dataDir ?? EnvironmentHelpers.GetDataDir(Path.Combine("WalletWasabi", "Backend"));
+			HostedServices = new HostedServices();
 		}
 
 		public Global() : this(null)
@@ -60,21 +59,9 @@ namespace WalletWasabi.Backend
 			// Make sure P2P works.
 			await InitializeP2pAsync(config.Network, config.GetBitcoinP2pEndPoint(), cancel);
 
-			// Initialize index building
-			var indexBuilderServiceDir = Path.Combine(DataDir, "IndexBuilderService");
-			var indexFilePath = Path.Combine(indexBuilderServiceDir, $"Index{RpcClient.Network}.dat");
-			var utxoSetFilePath = Path.Combine(indexBuilderServiceDir, $"UtxoSet{RpcClient.Network}.dat");
-			IndexBuilderService = new IndexBuilderService(RpcClient, BlockNotifier, indexFilePath, utxoSetFilePath);
-			Coordinator = new Coordinator(RpcClient.Network, BlockNotifier, Path.Combine(DataDir, "CcjCoordinator"), RpcClient, roundConfig);
-			IndexBuilderService.Synchronize();
-			Logger.LogInfo($"{nameof(IndexBuilderService)} is successfully initialized and started synchronization.");
-
-			await Coordinator.MakeSureTwoRunningRoundsAsync();
-			Logger.LogInfo("Chaumian CoinJoin Coordinator is successfully initialized and started two new rounds.");
-
 			if (roundConfig.FilePath != null)
 			{
-				RoundConfigWatcher = new ConfigWatcher(
+				HostedServices.Register(new ConfigWatcher(
 					TimeSpan.FromSeconds(10), // Every 10 seconds check the config
 					RoundConfig,
 					async () =>
@@ -89,10 +76,23 @@ namespace WalletWasabi.Backend
 						{
 							Logger.LogDebug(ex);
 						}
-					});
-				RoundConfigWatcher.Start();
-				Logger.LogInfo($"{nameof(RoundConfigWatcher)} is successfully started.");
+					}), "Config Watcher");
 			}
+
+			await HostedServices.StartAllAsync(cancel);
+
+			// Initialize index building
+			var indexBuilderServiceDir = Path.Combine(DataDir, "IndexBuilderService");
+			var indexFilePath = Path.Combine(indexBuilderServiceDir, $"Index{RpcClient.Network}.dat");
+			var utxoSetFilePath = Path.Combine(indexBuilderServiceDir, $"UtxoSet{RpcClient.Network}.dat");
+			var blockNotifier = HostedServices.FirstOrDefault<BlockNotifier>();
+			IndexBuilderService = new IndexBuilderService(RpcClient, blockNotifier, indexFilePath, utxoSetFilePath);
+			Coordinator = new Coordinator(RpcClient.Network, blockNotifier, Path.Combine(DataDir, "CcjCoordinator"), RpcClient, roundConfig);
+			IndexBuilderService.Synchronize();
+			Logger.LogInfo($"{nameof(IndexBuilderService)} is successfully initialized and started synchronization.");
+
+			await Coordinator.MakeSureTwoRunningRoundsAsync();
+			Logger.LogInfo("Chaumian CoinJoin Coordinator is successfully initialized and started two new rounds.");
 		}
 
 		private async Task InitializeP2pAsync(Network network, EndPoint endPoint, CancellationToken cancel)
@@ -103,8 +103,7 @@ namespace WalletWasabi.Backend
 			// We have to find it, because it's cloned by the node and not perfectly cloned (event handlers cannot be cloned.)
 			P2pNode = new P2pNode(network, endPoint, new MempoolService(), $"/WasabiCoordinator:{Constants.BackendMajorVersion.ToString()}/");
 			await P2pNode.ConnectAsync(cancel).ConfigureAwait(false);
-			BlockNotifier = new BlockNotifier(TimeSpan.FromSeconds(7), RpcClient, P2pNode);
-			BlockNotifier.Start();
+			HostedServices.Register(new BlockNotifier(TimeSpan.FromSeconds(7), new RpcWrappedClient(RpcClient), P2pNode), "Block Notifier");
 		}
 
 		private async Task AssertRpcNodeFullyInitializedAsync()
