@@ -1,5 +1,4 @@
 using Avalonia;
-using Avalonia.Input.Platform;
 using Avalonia.Threading;
 using ReactiveUI;
 using System;
@@ -16,6 +15,7 @@ using WalletWasabi.Blockchain.Keys;
 using WalletWasabi.Gui.Helpers;
 using WalletWasabi.Gui.Tabs.WalletManager;
 using WalletWasabi.Gui.ViewModels;
+using WalletWasabi.Logging;
 using WalletWasabi.Hwi;
 using WalletWasabi.Hwi.Exceptions;
 
@@ -37,10 +37,9 @@ namespace WalletWasabi.Gui.Controls.WalletExplorer
 		public ReactiveCommand<Unit, Unit> CopyLabel { get; }
 		public ReactiveCommand<Unit, Unit> ToggleQrCode { get; }
 		public ReactiveCommand<Unit, Unit> ChangeLabelCommand { get; }
-
 		public ReactiveCommand<Unit, Unit> DisplayAddressOnHwCommand { get; }
-
 		public ReactiveCommand<Unit, Unit> GenerateCommand { get; }
+		public ReactiveCommand<Unit, Unit> SaveQRCodeCommand { get; }
 
 		public ReceiveTabViewModel(WalletViewModel walletViewModel)
 			: base("Receive", walletViewModel)
@@ -93,7 +92,7 @@ namespace WalletWasabi.Gui.Controls.WalletExplorer
 					await address.TryCopyToClipboardAsync();
 				});
 
-			var isCoinListItemSelected = this.WhenAnyValue(x => x.SelectedAddress).Select(coin => coin != null);
+			var isCoinListItemSelected = this.WhenAnyValue(x => x.SelectedAddress).Select(coin => coin is { });
 
 			CopyAddress = ReactiveCommand.CreateFromTask(async () =>
 			{
@@ -106,27 +105,9 @@ namespace WalletWasabi.Gui.Controls.WalletExplorer
 			},
 			isCoinListItemSelected);
 
-			CopyLabel = ReactiveCommand.CreateFromTask(async () =>
-			{
-				try
-				{
-					await Application.Current.Clipboard.SetTextAsync(SelectedAddress.Label ?? string.Empty);
-				}
-				catch (Exception)
-				{ }
-			},
-			isCoinListItemSelected);
+			CopyLabel = ReactiveCommand.CreateFromTask(async () => await Application.Current.Clipboard.SetTextAsync(SelectedAddress.Label ?? string.Empty), isCoinListItemSelected);
 
-			ToggleQrCode = ReactiveCommand.Create(() =>
-				{
-					try
-					{
-						SelectedAddress.IsExpanded = !SelectedAddress.IsExpanded;
-					}
-					catch (Exception)
-					{ }
-				},
-				isCoinListItemSelected);
+			ToggleQrCode = ReactiveCommand.Create(() => ToggleSelectedAddress(), isCoinListItemSelected);
 
 #pragma warning disable IDE0053 // Use expression body for lambda expressions
 			ChangeLabelCommand = ReactiveCommand.Create(() => { SelectedAddress.InEditMode = true; });
@@ -147,7 +128,31 @@ namespace WalletWasabi.Gui.Controls.WalletExplorer
 				}
 			});
 
+			SaveQRCodeCommand = ReactiveCommand.CreateFromTask(async () =>
+			{
+				if (SelectedAddress is { })
+				{
+					await SelectedAddress.SaveQRCodeAsync();
+				}
+			});
+
+			Observable
+				.Merge(DisplayAddressOnHwCommand.ThrownExceptions)
+				.Merge(ChangeLabelCommand.ThrownExceptions)
+				.Merge(ToggleQrCode.ThrownExceptions)
+				.Merge(CopyAddress.ThrownExceptions)
+				.Merge(CopyLabel.ThrownExceptions)
+				.Merge(GenerateCommand.ThrownExceptions)
+				.Merge(SaveQRCodeCommand.ThrownExceptions)
+				.ObserveOn(RxApp.MainThreadScheduler)
+				.Subscribe(ex => NotificationHelpers.Error(ex.ToTypeMessageString()));
+
 			_suggestions = new ObservableCollection<SuggestionViewModel>();
+		}
+
+		private void ToggleSelectedAddress()
+		{
+			SelectedAddress.IsExpanded = !SelectedAddress.IsExpanded;
 		}
 
 		public override void OnOpen()
@@ -156,19 +161,10 @@ namespace WalletWasabi.Gui.Controls.WalletExplorer
 
 			Disposables = Disposables is null ? new CompositeDisposable() : throw new NotSupportedException($"Cannot open {GetType().Name} before closing it.");
 
-			Observable.FromEventPattern(Global.WalletService.TransactionProcessor,
-				nameof(Global.WalletService.TransactionProcessor.CoinReceived))
+			Observable
+				.FromEventPattern(Global.WalletService.TransactionProcessor, nameof(Global.WalletService.TransactionProcessor.CoinReceived))
 				.ObserveOn(RxApp.MainThreadScheduler)
 				.Subscribe(_ => InitializeAddresses())
-				.DisposeWith(Disposables);
-
-			Observable.Merge(DisplayAddressOnHwCommand.ThrownExceptions)
-				.Merge(ChangeLabelCommand.ThrownExceptions)
-				.Merge(ToggleQrCode.ThrownExceptions)
-				.Merge(CopyAddress.ThrownExceptions)
-				.Merge(CopyLabel.ThrownExceptions)
-				.ObserveOn(RxApp.MainThreadScheduler)
-				.Subscribe(ex => NotificationHelpers.Error(ex.ToTypeMessageString()))
 				.DisposeWith(Disposables);
 		}
 
@@ -191,11 +187,8 @@ namespace WalletWasabi.Gui.Controls.WalletExplorer
 				return;
 			}
 
-			foreach (HdPubKey key in walletService.KeyManager.GetKeys(x =>
-																		!x.Label.IsEmpty
-																		&& !x.IsInternal
-																		&& x.KeyState == KeyState.Clean)
-																	.Reverse())
+			IEnumerable<HdPubKey> keys = walletService.KeyManager.GetKeys(x => !x.Label.IsEmpty && !x.IsInternal && x.KeyState == KeyState.Clean).Reverse();
+			foreach (HdPubKey key in keys)
 			{
 				_addresses.Add(new AddressViewModel(key, Global));
 			}
@@ -270,7 +263,7 @@ namespace WalletWasabi.Gui.Controls.WalletExplorer
 			}
 			else
 			{
-				words[words.Length - 1] = word;
+				words[^1] = word;
 				Label = string.Join(", ", words) + ", ";
 			}
 
