@@ -37,8 +37,8 @@ namespace WalletWasabi.Gui.ViewModels
 
 		private bool UseTor { get; set; }
 
-		private ObservableAsPropertyHelper<RpcStatus> _bitcoinCoreStatus;
-		private ObservableAsPropertyHelper<UpdateStatus> _updateStatus;
+		private RpcStatus _bitcoinCoreStatus;
+		private UpdateStatus _updateStatus;
 		private bool _updateAvailable;
 		private bool _criticalUpdateAvailable;
 
@@ -64,13 +64,21 @@ namespace WalletWasabi.Gui.ViewModels
 			ActiveStatuses = new StatusBarStatusSet();
 		}
 
-		public void Initialize(NodesCollection nodes, WasabiSynchronizer synchronizer, UpdateChecker updateChecker)
+		public void Initialize(NodesCollection nodes, WasabiSynchronizer synchronizer)
 		{
 			Nodes = nodes;
 			Synchronizer = synchronizer;
 			HashChain = synchronizer.BitcoinStore.HashChain;
 			UseTor = Global.Config.UseTor; // Do not make it dynamic, because if you change this config settings only next time will it activate.
 			UseBitcoinCore = Global.Config.StartLocalBitcoinCoreOnStartup;
+			var hostedServices = Global.HostedServices;
+
+			var updateChecker = hostedServices.FirstOrDefault<UpdateChecker>();
+			Guard.NotNull(nameof(updateChecker), updateChecker);
+			UpdateStatus = updateChecker.UpdateStatus;
+
+			var rpcMonitor = hostedServices.FirstOrDefault<RpcMonitor>();
+			BitcoinCoreStatus = rpcMonitor?.RpcStatus ?? RpcStatus.Unresponsive;
 
 			_status = ActiveStatuses.WhenAnyValue(x => x.CurrentStatus)
 				.ObserveOn(RxApp.MainThreadScheduler)
@@ -113,19 +121,18 @@ namespace WalletWasabi.Gui.ViewModels
 				.Subscribe(usd => BtcPrice = $"${(long)usd}")
 				.DisposeWith(Disposables);
 
-			_bitcoinCoreStatus = Global.RpcMonitor
-					.WhenAnyValue(x => x.Status)
-					.Throttle(TimeSpan.FromMilliseconds(100))
+			if (rpcMonitor is { })
+			{
+				Observable.FromEventPattern<RpcStatus>(rpcMonitor, nameof(rpcMonitor.RpcStatusChanged))
 					.ObserveOn(RxApp.MainThreadScheduler)
-					.ToProperty(this, x => x.BitcoinCoreStatus)
+					.Subscribe(e => BitcoinCoreStatus = e.EventArgs)
 					.DisposeWith(Disposables);
+			}
 
-			_updateStatus = Global.UpdateChecker
-					.WhenAnyValue(x => x.Status)
-					.Throttle(TimeSpan.FromMilliseconds(100))
-					.ObserveOn(RxApp.MainThreadScheduler)
-					.ToProperty(this, x => x.UpdateStatus)
-					.DisposeWith(Disposables);
+			Observable.FromEventPattern<UpdateStatus>(updateChecker, nameof(updateChecker.UpdateStatusChanged))
+				.ObserveOn(RxApp.MainThreadScheduler)
+				.Subscribe(e => UpdateStatus = e.EventArgs)
+				.DisposeWith(Disposables);
 
 			Observable.FromEventPattern<bool>(Synchronizer, nameof(Synchronizer.ResponseArrivedIsGenSocksServFail))
 				.ObserveOn(RxApp.MainThreadScheduler)
@@ -202,8 +209,6 @@ namespace WalletWasabi.Gui.ViewModels
 			}, this.WhenAnyValue(x => x.UpdateAvailable, x => x.CriticalUpdateAvailable, (x, y) => x || y));
 
 			this.RaisePropertyChanged(nameof(UpdateCommand)); // The binding happens after the constructor. So, if the command is not in constructor, then we need this line.
-
-			updateChecker.Start();
 		}
 
 		public ReactiveCommand<Unit, Unit> UpdateCommand { get; set; }
@@ -233,8 +238,18 @@ namespace WalletWasabi.Gui.ViewModels
 		}
 
 		public int FiltersLeft => _filtersLeft?.Value ?? 0;
-		public RpcStatus BitcoinCoreStatus => _bitcoinCoreStatus?.Value ?? RpcStatus.Unresponsive;
-		public UpdateStatus UpdateStatus => _updateStatus?.Value ?? new UpdateStatus(true, true);
+
+		public RpcStatus BitcoinCoreStatus
+		{
+			get => _bitcoinCoreStatus;
+			set => this.RaiseAndSetIfChanged(ref _bitcoinCoreStatus, value);
+		}
+
+		public UpdateStatus UpdateStatus
+		{
+			get => _updateStatus;
+			set => this.RaiseAndSetIfChanged(ref _updateStatus, value);
+		}
 
 		public bool UpdateAvailable
 		{
