@@ -10,19 +10,21 @@ using WalletWasabi.BitcoinCore;
 using WalletWasabi.BitcoinCore.Endpointing;
 using WalletWasabi.Blockchain.Mempool;
 using WalletWasabi.Helpers;
+using WalletWasabi.Services;
 using WalletWasabi.Stores;
 
 namespace WalletWasabi.Tests.Helpers
 {
 	public static class TestNodeBuilder
 	{
-		public static async Task<CoreNode> CreateAsync([CallerFilePath]string callerFilePath = null, [CallerMemberName]string callerMemberName = null, string additionalFolder = null, MempoolService mempoolService = null)
+		public static async Task<CoreNode> CreateAsync(HostedServices hostedServices, [CallerFilePath]string callerFilePath = null, [CallerMemberName]string callerMemberName = null, string additionalFolder = null, MempoolService mempoolService = null)
 		{
 			var network = Network.RegTest;
 			return await CoreNode.CreateAsync(
 				new CoreNodeParams(
 					network,
 					mempoolService ?? new MempoolService(),
+					hostedServices,
 					Path.Combine(Global.Instance.DataDir, EnvironmentHelpers.ExtractFileName(callerFilePath), callerMemberName, additionalFolder ?? ""),
 					tryRestart: true,
 					tryDeleteDataDir: true,
