@@ -5,6 +5,7 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using WalletWasabi.BitcoinCore;
+using WalletWasabi.Services;
 using WalletWasabi.Tests.Helpers;
 using Xunit;
 
@@ -15,14 +16,21 @@ namespace WalletWasabi.Tests.UnitTests.BitcoinCore
 		[Fact]
 		public async Task CanBuildCoreNodeAsync()
 		{
-			CoreNode coreNode = await TestNodeBuilder.CreateAsync();
+			using var services = new HostedServices();
+			var coreNode = await TestNodeBuilder.CreateAsync(services);
+			await services.StartAllAsync(CancellationToken.None);
 			await coreNode.TryStopAsync();
+			await services.StopAllAsync(CancellationToken.None);
 		}
 
 		[Fact]
 		public async Task NodesDifferAsync()
 		{
-			var coreNodes = await Task.WhenAll(TestNodeBuilder.CreateAsync(additionalFolder: "0"), TestNodeBuilder.CreateAsync(additionalFolder: "1"));
+			using var services1 = new HostedServices();
+			using var services2 = new HostedServices();
+			var coreNodes = await Task.WhenAll(TestNodeBuilder.CreateAsync(services1, additionalFolder: "0"), TestNodeBuilder.CreateAsync(services2, additionalFolder: "1"));
+			await services1.StartAllAsync(CancellationToken.None);
+			await services2.StartAllAsync(CancellationToken.None);
 			CoreNode node1 = coreNodes[0];
 			CoreNode node2 = coreNodes[1];
 			try
@@ -33,6 +41,8 @@ namespace WalletWasabi.Tests.UnitTests.BitcoinCore
 			}
 			finally
 			{
+				await services1.StopAllAsync(CancellationToken.None);
+				await services2.StopAllAsync(CancellationToken.None);
 				await Task.WhenAll(node1.TryStopAsync(), node2.TryStopAsync());
 			}
 		}
@@ -40,7 +50,9 @@ namespace WalletWasabi.Tests.UnitTests.BitcoinCore
 		[Fact]
 		public async Task RpcWorksAsync()
 		{
-			var coreNode = await TestNodeBuilder.CreateAsync();
+			using var services = new HostedServices();
+			var coreNode = await TestNodeBuilder.CreateAsync(services);
+			await services.StartAllAsync(CancellationToken.None);
 			try
 			{
 				var blockCount = await coreNode.RpcClient.GetBlockCountAsync();
@@ -48,6 +60,7 @@ namespace WalletWasabi.Tests.UnitTests.BitcoinCore
 			}
 			finally
 			{
+				await services.StopAllAsync(CancellationToken.None);
 				await coreNode.TryStopAsync();
 			}
 		}
@@ -55,7 +68,9 @@ namespace WalletWasabi.Tests.UnitTests.BitcoinCore
 		[Fact]
 		public async Task P2pWorksAsync()
 		{
-			var coreNode = await TestNodeBuilder.CreateAsync();
+			using var services = new HostedServices();
+			var coreNode = await TestNodeBuilder.CreateAsync(services);
+			await services.StartAllAsync(CancellationToken.None);
 			using var node = await coreNode.CreateNewP2pNodeAsync();
 			try
 			{
@@ -65,6 +80,7 @@ namespace WalletWasabi.Tests.UnitTests.BitcoinCore
 			}
 			finally
 			{
+				await services.StopAllAsync(CancellationToken.None);
 				node.Disconnect();
 				await coreNode.TryStopAsync();
 			}
