@@ -9,16 +9,18 @@ using WalletWasabi.Logging;
 
 namespace WalletWasabi.BitcoinCore.Monitoring
 {
-	public class P2pReconnector : PeriodicRunner<bool>
+	public class P2pReconnector : PeriodicRunner
 	{
 		public P2pNode P2pNode { get; set; }
+		private TaskCompletionSource<object> Success { get; }
 
-		public P2pReconnector(TimeSpan period, P2pNode p2pNode) : base(period, false)
+		public P2pReconnector(TimeSpan period, P2pNode p2pNode) : base(period)
 		{
 			P2pNode = Guard.NotNull(nameof(p2pNode), p2pNode);
+			Success = new TaskCompletionSource<object>();
 		}
 
-		protected override async Task<bool> ActionAsync(CancellationToken cancel)
+		protected override async Task ActionAsync(CancellationToken cancel)
 		{
 			try
 			{
@@ -33,16 +35,26 @@ namespace WalletWasabi.BitcoinCore.Monitoring
 			}
 
 			Logger.LogInfo("Successfully reconnected to P2P.");
-			return true;
+
+			Success.TrySetResult(null);
 		}
 
 		public async Task StartAndAwaitReconnectionAsync(CancellationToken cancel)
 		{
-			Stop?.Dispose();
-			Stop = CancellationTokenSource.CreateLinkedTokenSource(cancel);
-			ForeverTask = ForeverMethodAsync(x => x is true);
-			await ForeverTask.ConfigureAwait(false);
-			await StopAsync().ConfigureAwait(false);
+			await StartAsync(cancel).ConfigureAwait(false);
+			using var ctr = cancel.Register(() => Success.SetResult(null));
+			await Success.Task.ConfigureAwait(false);
+
+			try
+			{
+				using var cts = new CancellationTokenSource(TimeSpan.FromSeconds(21));
+				await StopAsync(cts.Token).ConfigureAwait(false);
+			}
+			catch (Exception ex)
+			{
+				Logger.LogError(ex);
+			}
+			Dispose();
 		}
 	}
 }
