@@ -11,26 +11,44 @@ using WalletWasabi.Helpers;
 
 namespace WalletWasabi.BitcoinCore.Monitoring
 {
-	public class RpcFeeProvider : PeriodicRunner<AllFeeEstimate>, IFeeProvider
+	public class RpcFeeProvider : PeriodicRunner, IFeeProvider
 	{
+		public event EventHandler<AllFeeEstimate> AllFeeEstimateChanged;
+
+		private AllFeeEstimate _allFeeEstimate;
+
+		public AllFeeEstimate AllFeeEstimate
+		{
+			get => _allFeeEstimate;
+			private set
+			{
+				if (value != _allFeeEstimate)
+				{
+					_allFeeEstimate = value;
+					AllFeeEstimateChanged?.Invoke(this, value);
+				}
+			}
+		}
+
 		public RPCClient RpcClient { get; set; }
 
-		public RpcFeeProvider(TimeSpan period, RPCClient rpcClient) : base(period, null)
+		public RpcFeeProvider(TimeSpan period, RPCClient rpcClient) : base(period)
 		{
 			RpcClient = Guard.NotNull(nameof(rpcClient), rpcClient);
 		}
 
-		protected override async Task<AllFeeEstimate> ActionAsync(CancellationToken cancel)
+		protected override async Task ActionAsync(CancellationToken cancel)
 		{
 			try
 			{
-				return await RpcClient.EstimateAllFeeAsync(EstimateSmartFeeMode.Conservative, true, true).ConfigureAwait(false);
+				var allFeeEstimate = await RpcClient.EstimateAllFeeAsync(EstimateSmartFeeMode.Conservative, true, true).ConfigureAwait(false);
+				AllFeeEstimate = allFeeEstimate;
 			}
 			catch
 			{
-				if (Status != null)
+				if (AllFeeEstimate != null)
 				{
-					Status = new AllFeeEstimate(Status, false);
+					AllFeeEstimate = new AllFeeEstimate(AllFeeEstimate, false);
 				}
 				throw;
 			}
