@@ -17,7 +17,7 @@ using WalletWasabi.Logging;
 
 namespace WalletWasabi.BitcoinCore
 {
-	public class P2pNode : IDisposable
+	public class P2pNode
 	{
 		private Node Node { get; set; }
 		private TrustedP2pBehavior TrustedP2pBehavior { get; set; }
@@ -77,6 +77,7 @@ namespace WalletWasabi.BitcoinCore
 			lock (SubscriptionLock)
 			{
 				TrustedP2pBehavior = Node.Behaviors.Find<TrustedP2pBehavior>();
+				Node.UncaughtException += Node_UncaughtException;
 				Node.StateChanged += P2pNode_StateChanged;
 				Node.Disconnected += Node_DisconnectedAsync;
 				TrustedP2pBehavior.BlockInv += TrustedP2pBehavior_BlockInv;
@@ -84,6 +85,13 @@ namespace WalletWasabi.BitcoinCore
 				MempoolService.TrustedNodeMode = Node.IsConnected;
 			}
 		}
+
+		private void Node_UncaughtException(Node sender, Exception ex)
+		{
+			Logger.LogInfo($"Node {sender.Peer.Endpoint} failed with exception: {ex}");
+		}
+
+		private Task ReconnectorTask { get; set; } = Task.CompletedTask;
 
 		private async void Node_DisconnectedAsync(Node node)
 		{
@@ -96,7 +104,8 @@ namespace WalletWasabi.BitcoinCore
 						return;
 					}
 					var reconnector = new P2pReconnector(TimeSpan.FromSeconds(7), this);
-					await reconnector.StartAndAwaitReconnectionAsync(Stop.Token).ConfigureAwait(false);
+					ReconnectorTask = reconnector.StartAndAwaitReconnectionAsync(Stop.Token);
+					await ReconnectorTask.ConfigureAwait(false);
 				}
 			}
 			catch (Exception ex)
@@ -123,31 +132,13 @@ namespace WalletWasabi.BitcoinCore
 			}
 		}
 
-		#region IDisposable Support
-
-		private volatile bool _disposedValue = false; // To detect redundant calls
-
-		protected virtual void Dispose(bool disposing)
+		public async Task DisposeAsync()
 		{
-			if (!_disposedValue)
-			{
-				if (disposing)
-				{
-					Stop?.Cancel();
-					Disconnect();
-					Stop?.Dispose();
-					Stop = null;
-				}
-
-				_disposedValue = true;
-			}
-		}
-
-		// This code added to correctly implement the disposable pattern.
-		public void Dispose()
-		{
-			// Do not change this code. Put cleanup code in Dispose(bool disposing) above.
-			Dispose(true);
+			Stop?.Cancel();
+			await ReconnectorTask.ConfigureAwait(false);
+			Disconnect();
+			Stop?.Dispose();
+			Stop = null;
 		}
 
 		/// <summary>
@@ -155,26 +146,29 @@ namespace WalletWasabi.BitcoinCore
 		/// </summary>
 		public void Disconnect()
 		{
-			if (Node is { })
+			Node node = Node;
+			if (node is { })
 			{
 				lock (SubscriptionLock)
 				{
 					MempoolService.TrustedNodeMode = false;
 					if (NodeEventsSubscribed)
 					{
-						if (TrustedP2pBehavior is { })
+						var trustedP2pBehavior = TrustedP2pBehavior;
+						if (trustedP2pBehavior is { })
 						{
-							TrustedP2pBehavior.BlockInv -= TrustedP2pBehavior_BlockInv;
+							trustedP2pBehavior.BlockInv -= TrustedP2pBehavior_BlockInv;
 						}
-						Node.Disconnected -= Node_DisconnectedAsync;
-						Node.StateChanged -= P2pNode_StateChanged;
+						node.Disconnected -= Node_DisconnectedAsync;
+						node.StateChanged -= P2pNode_StateChanged;
+						node.UncaughtException -= Node_UncaughtException;
 						NodeEventsSubscribed = false;
 					}
 				}
 
 				try
 				{
-					Node?.Disconnect();
+					node.Disconnect();
 				}
 				catch (Exception ex)
 				{
@@ -184,7 +178,7 @@ namespace WalletWasabi.BitcoinCore
 				{
 					try
 					{
-						Node?.Dispose();
+						node.Dispose();
 					}
 					catch (Exception ex)
 					{
@@ -198,7 +192,5 @@ namespace WalletWasabi.BitcoinCore
 				}
 			}
 		}
-
-		#endregion IDisposable Support
 	}
 }
