@@ -7,18 +7,19 @@ namespace Gma.QrCodeNet.Encoding.Positioning.Stencils
 		{
 		}
 
-		private static bool[,] PositionDetection { get; } = new[,]
-																		{
-																			{ O, O, O, O, O, O, O, O, O },
-																			{ O, X, X, X, X, X, X, X, O },
-																			{ O, X, O, O, O, O, O, X, O },
-																			{ O, X, O, X, X, X, O, X, O },
-																			{ O, X, O, X, X, X, O, X, O },
-																			{ O, X, O, X, X, X, O, X, O },
-																			{ O, X, O, O, O, O, O, X, O },
-																			{ O, X, X, X, X, X, X, X, O },
-																			{ O, O, O, O, O, O, O, O, O }
-																		};
+		private static bool[,] PositionDetection { get; } =
+			new[,]
+			{
+				{ O, O, O, O, O, O, O, O, O },
+				{ O, X, X, X, X, X, X, X, O },
+				{ O, X, O, O, O, O, O, X, O },
+				{ O, X, O, X, X, X, O, X, O },
+				{ O, X, O, X, X, X, O, X, O },
+				{ O, X, O, X, X, X, O, X, O },
+				{ O, X, O, O, O, O, O, X, O },
+				{ O, X, X, X, X, X, X, X, O },
+				{ O, O, O, O, O, O, O, O, O }
+			};
 
 		public override bool[,] Stencil => PositionDetection;
 
