@@ -1,4 +1,5 @@
 using System;
+using System.ComponentModel;
 using System.Threading;
 using System.Threading.Tasks;
 using WalletWasabi.Bases;
@@ -10,18 +11,36 @@ using WalletWasabi.WebClients.Wasabi;
 
 namespace WalletWasabi.Services
 {
-	public class UpdateChecker : PeriodicRunner<UpdateStatus>
+	public class UpdateChecker : PeriodicRunner
 	{
+		private UpdateStatus _updateStatus;
+
 		public WasabiClient WasabiClient { get; }
 
-		public UpdateChecker(TimeSpan period, WasabiClient client) : base(period, new UpdateStatus(true, true))
+		public UpdateStatus UpdateStatus
 		{
-			WasabiClient = Guard.NotNull(nameof(client), client);
+			get => _updateStatus;
+			private set
+			{
+				if (value != _updateStatus)
+				{
+					_updateStatus = value;
+					UpdateStatusChanged?.Invoke(this, value);
+				}
+			}
 		}
 
-		protected override async Task<UpdateStatus> ActionAsync(CancellationToken cancel)
+		public event EventHandler<UpdateStatus> UpdateStatusChanged;
+
+		public UpdateChecker(TimeSpan period, WasabiClient client) : base(period)
 		{
-			return await WasabiClient.CheckUpdatesAsync(cancel).ConfigureAwait(false);
+			WasabiClient = Guard.NotNull(nameof(client), client);
+			UpdateStatus = new UpdateStatus(true, true);
+		}
+
+		protected override async Task ActionAsync(CancellationToken cancel)
+		{
+			UpdateStatus = await WasabiClient.CheckUpdatesAsync(cancel).ConfigureAwait(false);
 		}
 	}
 }
